Creator (#42)
* organizer api controller

* logging flow

* run build for logging

* define output result of transaction raw

* result types

* fixing raw query

* update prisma package & adding sql generated to the prisma file

* fixing raw sql

+++ prisma/sql/getOrganizerStat.sql
@@ -1,7 +1,7 @@
 SELECT date_trunc($1, t."createdAt")::date AS date, 
-  SUM(t.total_amount) AS total_revenue, 
-  COUNT(t.transaction_id) AS total_transaction, 
-  SUM(td.sum_seat) AS total_seat 
+  SUM(t.total_amount)::numeric AS total_revenue, 
+  COUNT(t.transaction_id)::numeric AS total_transaction, 
+  SUM(td.sum_seat)::numeric AS total_seat 
 FROM "transaction" t 
 JOIN (
   SELECT td.transaction_id, SUM(td.quantity_bought) AS sum_seat
